# internal/docs/openapi.yaml
openapi: 3.1.0
info:
    title: ERM Dokter API
    version: 1.0.0
    description: |
        API Elektronik Rekam Medis (ERM) Dokter untuk Rumah Sakit & Klinik berbasis database Khanza.

        ### Fitur Utama:
        - **Authentication**: JWT Bearer Token dengan enkripsi password AES.
        - **Rawat Jalan**: Antrean pasien dokter, pencarian & filter, detail kunjungan pasien.
        - **Pemeriksaan Medis (SOAP & TTV)**: Riwayat pemeriksaan, pencatatan SOAP (Subjective, Objective, Assessment, Plan) baru, update pemeriksaan, dan penghapusan catatan dengan proteksi kepemilikan dokter & batas waktu edit rekam medis (48 jam).
        - **Resep Obat**: Riwayat peresepan obat pasien per kunjungan dan riwayat peresepan seluruh kunjungan berdasarkan No. RM (mendukung filter Ralan, Ranap, Semua, dan rentang tanggal).
        - **Security & Privacy**: Enkripsi identifier URL (`id_kunjungan`, `id_pasien`, `id_pemeriksaan`, `id_obat`) menggunakan AES-GCM URL-safe token untuk mencegah ID enumeration / scraping.
    contact:
        name: Tim Pengembang ERM Dokter
servers:
    - url: http://192.168.30.153:8082
      description: Development Server (LAN / 192.168.30.153)
    - url: http://localhost:8082
      description: Development Server (Localhost)
security:
    - BearerAuth: []
...
components:
    securitySchemes:
        BearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT
            description: Masukkan JWT Token yang didapatkan dari endpoint login (/api/v1/auth/login).
    schemas:
...
        MasterBerkasDigital:
            type: object
            properties:
                kode:
                    type: string
                    example: "015"
                nama:
                    type: string
                    example: "PATOLOGI ANATOMI"
...
paths:
...
    /api/v1/master/berkas-digital:
        get:
            summary: Master Kategori Berkas Digital
            description: Mengambil seluruh daftar referensi kategori berkas digital aktif di rumah sakit dari master_berkas_digital.
            tags:
                - Master Data
            security:
                - BearerAuth: []
            responses:
                "200":
                    description: Berhasil mengambil master berkas digital
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    code:
                                        type: integer
                                        example: 200
                                    status:
                                        type: string
                                        example: OK
                                    message:
                                        type: string
                                        example: Berhasil mengambil master berkas digital
                                    data:
                                        type: array
                                        items:
                                            $ref: "#/components/schemas/MasterBerkasDigital"
    /api/v1/berkas-digital/{id_berkas}:
        get:
            summary: Secure Stream Berkas Digital Universal (PDF / Gambar)
            description: Mengalirkan (reverse proxy stream) dokumen PDF atau gambar dari server penyimpanan Khanza secara aman untuk seluruh modul RS (Laboratorium, Radiologi, EKG, USG, dll.).
            tags:
                - Berkas Digital
            security:
                - BearerAuth: []
            parameters:
                - name: id_berkas
                  in: path
                  required: true
                  description: Token URL berkas terenkripsi
                  schema:
                      type: string
                      example: "bK8mQ1..."
            responses:
                "200":
                    description: File stream PDF / Gambar
                    content:
                        application/pdf:
                            schema:
                                type: string
                                format: binary
                        image/jpeg:
                            schema:
                                type: string
                                format: binary
                        image/png:
                            schema:
                                type: string
                                format: binary
                "400":
                    description: Token berkas tidak valid
                "404":
                    description: Berkas tidak ditemukan di server penyimpanan
...